requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.12.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
asyncpg==0.31.0
    # via agent-backend (pyproject.toml)
cachetools==6.2.4
    # via google-auth
certifi==2025.11.12
    # via
    #   httpcore
    #   httpx
    #   requests
    #   sentry-sdk
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via
    #   rich-toolkit
    #   typer
    #   uvicorn
dnspython==2.8.0
    # via email-validator
email-validator==2.3.0
    # via
    #   fastapi
    #   pydantic
fastapi==0.125.0
    # via agent-backend (pyproject.toml)
fastapi-cli==0.0.16
    # via fastapi
fastapi-cloud-cli==0.7.0
    # via fastapi-cli
fastar==0.8.0
    # via fastapi-cloud-cli
google-api-core==2.28.1
    # via
    #   google-api-python-client
    #   google-cloud-appengine-logging
    #   google-cloud-core
    #   google-cloud-logging
    #   google-cloud-secret-manager
google-api-python-client==2.187.0
    # via agent-backend (pyproject.toml)
google-auth==2.45.0
    # via
    #   agent-backend (pyproject.toml)
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-auth-oauthlib
    #   google-cloud-appengine-logging
    #   google-cloud-core
    #   google-cloud-logging
    #   google-cloud-secret-manager
google-auth-httplib2==0.3.0
    # via google-api-python-client
google-auth-oauthlib==1.2.2
    # via agent-backend (pyproject.toml)
google-cloud-appengine-logging==1.7.0
    # via google-cloud-logging
google-cloud-audit-log==0.4.0
    # via google-cloud-logging
google-cloud-core==2.5.0
    # via google-cloud-logging
google-cloud-logging==3.13.0
    # via agent-backend (pyproject.toml)
google-cloud-secret-manager==2.26.0
    # via agent-backend (pyproject.toml)
googleapis-common-protos==1.72.0
    # via
    #   google-api-core
    #   google-cloud-audit-log
    #   grpc-google-iam-v1
    #   grpcio-status
grpc-google-iam-v1==0.14.3
    # via
    #   google-cloud-logging
    #   google-cloud-secret-manager
grpcio==1.76.0
    # via
    #   google-api-core
    #   google-cloud-appengine-logging
    #   google-cloud-secret-manager
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
grpcio-status==1.76.0
    # via google-api-core
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httplib2==0.31.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
httptools==0.7.1
    # via uvicorn
httpx==0.28.1
    # via
    #   agent-backend (pyproject.toml)
    #   fastapi
    #   fastapi-cloud-cli
idna==3.11
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
importlib-metadata==8.7.1
    # via opentelemetry-api
jinja2==3.1.6
    # via fastapi
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.3
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
oauthlib==3.3.1
    # via requests-oauthlib
opentelemetry-api==1.39.1
    # via google-cloud-logging
proto-plus==1.27.0
    # via
    #   google-api-core
    #   google-cloud-appengine-logging
    #   google-cloud-logging
    #   google-cloud-secret-manager
protobuf==6.33.2
    # via
    #   google-api-core
    #   google-cloud-appengine-logging
    #   google-cloud-audit-log
    #   google-cloud-logging
    #   google-cloud-secret-manager
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
    #   proto-plus
psycopg2-binary==2.9.11
    # via agent-backend (pyproject.toml)
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.12.5
    # via
    #   fastapi
    #   fastapi-cloud-cli
    #   sqlmodel
pydantic-core==2.41.5
    # via pydantic
pygments==2.19.2
    # via rich
pyparsing==3.2.5
    # via httplib2
python-dotenv==1.2.1
    # via
    #   agent-backend (pyproject.toml)
    #   uvicorn
python-json-logger==4.0.0
    # via agent-backend (pyproject.toml)
python-multipart==0.0.21
    # via fastapi
pyyaml==6.0.3
    # via uvicorn
redis==7.1.0
    # via agent-backend (pyproject.toml)
requests==2.32.5
    # via
    #   google-api-core
    #   requests-oauthlib
requests-oauthlib==2.0.0
    # via google-auth-oauthlib
rich==14.2.0
    # via
    #   rich-toolkit
    #   typer
rich-toolkit==0.17.1
    # via
    #   fastapi-cli
    #   fastapi-cloud-cli
rignore==0.7.6
    # via fastapi-cloud-cli
rsa==4.9.1
    # via google-auth
sentry-sdk==2.48.0
    # via fastapi-cloud-cli
shellingham==1.5.4
    # via typer
sqlalchemy==2.0.45
    # via sqlmodel
sqlmodel==0.0.27
    # via agent-backend (pyproject.toml)
starlette==0.50.0
    # via fastapi
typer==0.20.0
    # via
    #   fastapi-cli
    #   fastapi-cloud-cli
typing-extensions==4.15.0
    # via
    #   anyio
    #   fastapi
    #   grpcio
    #   opentelemetry-api
    #   pydantic
    #   pydantic-core
    #   rich-toolkit
    #   sqlalchemy
    #   starlette
    #   typer
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
uritemplate==4.2.0
    # via google-api-python-client
urllib3==2.6.2
    # via
    #   requests
    #   sentry-sdk
uvicorn==0.38.0
    # via
    #   fastapi
    #   fastapi-cli
    #   fastapi-cloud-cli
uvloop==0.22.1
    # via uvicorn
watchfiles==1.1.1
    # via uvicorn
websockets==15.0.1
    # via uvicorn
zipp==3.23.0
    # via importlib-metadata
